6_sem/security/lab1/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psutil==5.9.4
python-dotenv==1.0.0
Werkzeug==2.2.3
